+++ persistent-sqlite/ChangeLog.md
@@ -1,3 +1,8 @@
+## 2.6.0.2
+
+* Added functions to monitor (status) and control (softHeapLimit) process-wide
+  SQLite memory usage.
+
 ## 2.6.0.1
 
 * Ensure connection is closed if wrapConnectionWal fails